[chore 158382565] Refactor folder structure.
