+++ api/__init__.py
@@ -13,16 +13,16 @@ def create_app(config_name):
     app.config.from_pyfile('config.py')
     app.secret_key = os.getenv('SECRET_KEY')
 
-    from api.auth.views import auth
+    from api.auth.auth import auth
     app.register_blueprint(auth)
 
-    from api.mrequests.views import mrequests
+    from api.mrequests.mrequests import mrequests
     app.register_blueprint(mrequests)
 
-    from api.mrequests.admin import admin_bp
+    from api.admin.admin import admin_bp
     app.register_blueprint(admin_bp)
 
-    from api.mrequests.super_admin import super_admin_bp
+    from api.super_admin.super_admin import super_admin_bp
     app.register_blueprint(super_admin_bp)
 
     return app
