+++ api/mrequests/__init__.py
@@ -3,7 +3,5 @@
 from flask import Blueprint
 
 mrequests = Blueprint('mrequests', __name__, url_prefix='/api/v2/users/requests')
-admin_bp = Blueprint('admin', __name__, url_prefix='/api/v2/requests')
-super_admin_bp =  Blueprint('super_admin', __name__, url_prefix='/api/v2/admin')
 
-from . import views, admin, super_admin
+from . import mrequests
